correciones js ver activos en plan, reporte agregando titulo, validar al eliminar control

+++ protected/controllers/ControlController.php
@@ -108,12 +108,22 @@ class ControlController extends Controller
     public function actionDelete($id)
     {
         if (Yii::app()->request->isPostRequest) {
-// we only allow deletion via POST request
-            $this->loadModel($id)->delete();
-
-// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-            if (!isset($_GET['ajax']))
+            try{
+                $grupo_activo = AnalisisControl::model()->findByAttributes(['control_id'=>$id]);
+                if(!is_null($grupo_activo)){
+                    throw new Exception("Error. Este control ya esta asociado a un analisis.");
+                }
+                $this->loadModel($id)->delete();
+                $data = "Se elimino correctamente el control";
+                echo CJSON::encode($data);
+            }catch (Exception $exception){
+                $data = $exception->getMessage();
+                echo CJSON::encode($data);
+                die();
+            }
+            if (!isset($_GET['ajax'])) {
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+            }
         } else
             throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
